Add filtering to favorite component
Added filtering so that when the icon is a yellow star that note will appear in the favorite page and when it is just a star border or unclicked then it will not appear there. Now we have to fix the problem that if you delete the note or that note gets expired then your favorited or saved notes will be gone from your favorite page. Progresses issue #1